% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_process_fuzzy_module.R
\name{assure_var_id}
\alias{assure_var_id}
\title{Assure that var_id is the same for the current outputfuzzymatch_df
and the aux_fuzzy_status that is read in}
\usage{
assure_var_id(output_fuzzymatch_df, aux_fuzzy_status)
}
\arguments{
\item{output_fuzzymatch_df}{A data.frame with the newly generated fuzzymatches}

\item{aux_fuzzy_status}{A data.frame containing the fuzzymatches with is_truematch manually checked}
}
\description{
Assure that var_id is the same for the current outputfuzzymatch_df
and the aux_fuzzy_status that is read in
}
