% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{bind_fuzzymatches}
\alias{bind_fuzzymatches}
\title{Bind the status of fuzzymatches with the fuzzymatches created in this session
so that the two stay in sync even if the fuzzymatch function is changed
Note: this function may have to be updated to handle multiple auxiliary fuzzymatch files}
\usage{
bind_fuzzymatches(output_fuzzymatch_df, aux_fuzzy_status)
}
\arguments{
\item{output_fuzzymatch_df}{A data.frame with the newly generated fuzzymatches}

\item{aux_fuzzy_status}{A data.frame containing the fuzzymatches with is_truematch manually checked}
}
\description{
Bind the status of fuzzymatches with the fuzzymatches created in this session
so that the two stay in sync even if the fuzzymatch function is changed
Note: this function may have to be updated to handle multiple auxiliary fuzzymatch files
}
\keyword{internal}
