% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{check.anymatch}
\alias{check.anymatch}
\title{Check for any matches}
\usage{
check.anymatch(df, group, checkfor = NULL, match_type)
}
\arguments{
\item{df}{A data.frame with variety names matched to database names}

\item{group}{A column name by which to group and search for a match
(e.g. to determine if there are any matches for the variety column, use: variety)}

\item{checkfor}{A column name to check whether there is a match
(This parameter is used only for exact matching.  For fuzzymatches, the status of the match
must be manually entered)}

\item{match_type}{A string that is either 'db' or 'raw', which refers to processesing
database entries (only one entry per cultivar) or raw entries which may have many
alternate spellings for one cultivar}
}
\description{
After a variety name match (exact or fuzzymatch), check whether there are any
exact or approved matches.
}
