% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{check.matches}
\alias{check.matches}
\title{Check that matches have one database match}
\usage{
check.matches(match_df)
}
\arguments{
\item{match_df}{A data.frame with variety names and database matches
(including volumns var_id and db_id)}
}
\description{
Check that matches have one database match
}
