% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{clean_manualfuzzy}
\alias{clean_manualfuzzy}
\title{Clean the file of fuzzy matches with manual checks}
\usage{
clean_manualfuzzy(match_type = "raw", df)
}
\arguments{
\item{match_type}{A string that is either 'db' or 'raw', which refers to processesing
database entries (only one entry per cultivar) or raw entries which may have many
alternate spellings for one cultivar}

\item{df}{A data.frame, which can be provided in lieu of the curation_folder
and filename}

\item{curation_folder}{A path to the curation folder}
}
\description{
Fuzzy matching outputs a file of potential/fuzzy matches.
A user must go in a assign whether the fuzzy match is a true match.
}
\keyword{internal}
