% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_final_matches_module.R
\name{collect_final_matches}
\alias{collect_final_matches}
\title{Module to collect all the matches from each stage of the variety matching process}
\usage{
collect_final_matches(match_list, names_raw, knitroutputfolder)
}
\arguments{
\item{match_list}{A list of the output from each step of the variety matching process}

\item{names_raw}{A data.frame of the names created by create_intid()}

\item{knitroutputfolder}{A path to the output location}

\item{csv_suffix}{A string denoting the suffix for the .csv name}
}
\description{
Note: If no match has been found, then the returned match == the original name
}
\seealso{
Other match variety functions: 
\code{\link{create_intid}()}
}
\concept{match variety functions}
