% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{collect_matches}
\alias{collect_matches}
\title{Combine all the matches from all of the cultivar matching steps}
\usage{
collect_matches(ls, ...)
}
\arguments{
\item{ls}{A list of results created with list2 from any of the cultivar matching steps}

\item{...}{Any number of bare variables to be retained in the output of the matches}
}
\description{
Combine all the matches from all of the cultivar matching steps
}
\keyword{internal}
