% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.rules.R
\name{create.rules}
\alias{create.rules}
\title{Create rules using validator function from the validate package}
\usage{
create.rules(df_type, db_folder)
}
\arguments{
\item{df_type}{A string "trial_data" or "trials_metadata" denoting the type of validation
rules to generate}

\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
This functions reads in codebooks_all_db.csv and generates rules from the
information provided in this main codebook
}
\details{
Note: grepl returns FALSE if the tested object is NA,
so NAs are not correctly detected in grepl checks.
Currently, the validator package does not accept str_detect.
As a workaround, for date fields, an extra check for NAs is implemented.
}
