% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dm.R
\name{create_dm}
\alias{create_dm}
\title{Create a data model object from the main codebook}
\usage{
create_dm(codebook_main_path)
}
\arguments{
\item{codebook_main_path}{A string that is a path to the main codebook}
}
\description{
This function creates a data model from the codebook to show
the relationship between the controlled variables (e.g. where
the controlled vocabularies are defined)
}
