% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{create_names_nomatch}
\alias{create_names_nomatch}
\title{Create a data.frame of the names of varieties with no match}
\usage{
create_names_nomatch(df_nomatch, ...)
}
\arguments{
\item{df_nomatch}{A data.frame of varieties without cultivar matches
from the database.  This will generally be the output of the function
clean_manualfuzzy().}

\item{...}{Extra columns to select}
}
\description{
Create a data.frame of the names of varieties with no match
}
\keyword{internal}
