% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_create_new_names_module.R
\name{create_new_db_entries}
\alias{create_new_db_entries}
\title{Need to standardize crop type before the curation process
This function is based on create_newdbnames() and create_newdbnames_db()}
\usage{
create_new_db_entries(
  auxiliary_files,
  filename,
  matches_temp_list,
  data_intid,
  match_type
)
}
\description{
Need to standardize crop type before the curation process
This function is based on create_newdbnames() and create_newdbnames_db()
}
