% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_process_fuzzy_module.R
\name{deprecated_process_fuzzymatch}
\alias{deprecated_process_fuzzymatch}
\title{Module to process fuzzymatches given an auxiliary file}
\usage{
deprecated_process_fuzzymatch(output_fuzzymatch_df, aux_fuzzy_status)
}
\arguments{
\item{output_fuzzymatch_df}{A data.frame with the newly generated fuzzymatches}

\item{aux_fuzzy_status}{A data.frame containing the fuzzymatches with is_truematch manually checked}
}
\description{
Module to process fuzzymatches given an auxiliary file
}
\keyword{internal}
