% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fuzzy_module.R
\name{extract_trail_digits}
\alias{extract_trail_digits}
\title{Extract the trailing digits from a column of type character and
return the digits in a new column (digits are returned as characters)}
\usage{
extract_trail_digits(df, x, rm_single)
}
\arguments{
\item{df}{A data.frame}

\item{x}{The bare variable name of the column to extract digits}
}
\description{
Extract the trailing digits from a column of type character and
return the digits in a new column (digits are returned as characters)
}
