% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_col_info.R
\name{find_col_info}
\alias{find_col_info}
\title{Find column information}
\usage{
find_col_info(df, cols_check, by_col)
}
\arguments{
\item{df}{A data.frame to summarise}

\item{cols_check}{A character vector of column names to check}

\item{by_col}{A bare variable to summarise}
}
\description{
For a data.frame, loop through the columns provided (cols_check)
and summarise the number of non-NA entries and the groups that
contain non-NA entries (for example if want to search trial, year, or filename),
and return the first non-NA entry as an example
}
\seealso{
Other validation functions: 
\code{\link{confront_data}()},
\code{\link{create_rules}()},
\code{\link{validate_colnames}()}
}
\concept{validation functions}
