% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R
\name{get.variety_db}
\alias{get.variety_db}
\title{Get cultivar names from the database for barley and wheat,
along with unique identifier (no spaces or special characters)
Filter the cultivar list by the date that the cultivar was added.
The default value is "2021-01-01" because the original version of this function
only read in the original set of cultivars, which were added by J. Piaskowski
in 2020.
If select_before = NULL, then all dates will be returned.  Recommended usage
for curation is to use the current date, so that the curation is reproducible
even as the cultivar list is being continuously updated.}
\usage{
get.variety_db(
  db_folder,
  select_before = "2021-01-01",
  select_crops = NULL,
  for_matching = FALSE
)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{select_before}{A string in the format of Ymd.  The function returns
cultivars that were added to the datebase before this specified date.}

\item{select_crops}{A regular expression of crop types separated by |}
}
\description{
Get cultivar names from the database for barley and wheat,
along with unique identifier (no spaces or special characters)
Filter the cultivar list by the date that the cultivar was added.
The default value is "2021-01-01" because the original version of this function
only read in the original set of cultivars, which were added by J. Piaskowski
in 2020.
If select_before = NULL, then all dates will be returned.  Recommended usage
for curation is to use the current date, so that the curation is reproducible
even as the cultivar list is being continuously updated.
}
\seealso{
Other access codebook functions: 
\code{\link{list.db_books}()},
\code{\link{list.db_var}()},
\code{\link{readin.db}()}
}
\concept{access codebook functions}
