% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R
\name{get_col_index}
\alias{get_col_index}
\title{Get column index for codebook variables in a list of data frames}
\usage{
get_col_index(
  datalist,
  db_folder,
  codebook_name,
  required_only = TRUE,
  crop_types = NULL
)
}
\arguments{
\item{datalist}{A list of dataframes with columns that match codebook variables}

\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{codebook_name}{A string denoting the book name.
Use \code{list_db_books()} to see options}

\item{required_only}{A logical, FALSE returns all columns; TRUE returns required columns only}

\item{crop_types}{A vector containing the crop_types.
This is used to select the appropriate traits if codebook_name == "trial_data"}
}
\value{
Returns a list of integer vectors with the same list names as \code{datalist}.
}
\description{
This function is used to get the column index numbers for selected columns. The index vectors are used to set column styling for Excel data templates, i.e. highlighting required columns. Invalid codebook_names will throw an error.
}
\seealso{
Other access codebook functions: 
\code{\link{get_variety_db}()},
\code{\link{list_db_books}()},
\code{\link{list_db_var}()},
\code{\link{readin_db}()}
}
\concept{access codebook functions}
