% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_exact_module.R
\name{get_cultivar_rename}
\alias{get_cultivar_rename}
\title{Get cultivar rename list to match.  The cv_rename.csv is located
in the controlled vocabularies.}
\usage{
get_cultivar_rename(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
Get cultivar rename list to match.  The cv_rename.csv is located
in the controlled vocabularies.
}
