% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R,
%   R/factcurator_wrappers.R
\name{get_variety_db}
\alias{get_variety_db}
\alias{get.variety_db}
\title{Return cultivar names along with unique internal identifier (no spaces or special characters)}
\usage{
get_variety_db(
  db_folder,
  select_before = Sys.Date(),
  select_crops = NULL,
  for_matching = FALSE
)

get.variety_db(
  db_folder,
  select_before = "2021-01-01",
  select_crops = NULL,
  for_matching = FALSE
)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{select_before}{A string in the format of Ymd.  The function returns
cultivars that were added to the datebase before this specified date.}

\item{select_crops}{A regular expression of crops separated by |. Note that this regex
will filter on the \code{crop} column rather than \code{crop_type}. For example, there
are entries in the wheat file with crop Triticale, so to capture both the input
should be \code{select_crops = "wheat|triticale"}.}

\item{for_matching}{logical, whether the result will be used for matching. Default is \code{FALSE}.}
}
\description{
Filter the cultivar list by the date that the cultivar was added.
The default value selects before current date.
If select_before = NULL, then all dates will be returned.  Recommended usage
for curation is to use the current date, so that the curation is reproducible
even as the cultivar list is being continuously updated.
}
\details{
Cultivar names are stored by crop_type, this function will return all the requested
crop_types.
}
\seealso{
Other access codebook functions: 
\code{\link{list_db_books}()},
\code{\link{list_db_var}()},
\code{\link{readin_db}()}
}
\concept{access codebook functions}
