% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fuzzy_module.R
\name{is_string_overlap}
\alias{is_string_overlap}
\title{Check for overlap in strings
This function takes a dataframe and compares two columns based on whether
the strings overlap.  TRUE is returned if there is an overlap in either direction
(it doesn't matter which string is longer)}
\usage{
is_string_overlap(df, x1, x2)
}
\arguments{
\item{df}{A data.frame}

\item{x1}{The bare variable name of the column to compare}

\item{x2}{The bare variable name of the column to compare}
}
\description{
Check for overlap in strings
This function takes a dataframe and compares two columns based on whether
the strings overlap.  TRUE is returned if there is an overlap in either direction
(it doesn't matter which string is longer)
}
\keyword{internal}
