% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R
\name{list.db_books}
\alias{list.db_books}
\title{List the names of the codebooks in the database}
\usage{
list.db_books(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
List the names of the codebooks in the database
}
