% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R,
%   R/factcurator_wrappers.R
\name{list_db_books}
\alias{list_db_books}
\alias{list.db_books}
\title{List the names of the codebooks in the database}
\usage{
list_db_books(db_folder)

list.db_books(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
List the names of the codebooks in the database

A wrapper for list_db_books to maintain backwards compatibility
}
\seealso{
Other access codebook functions: 
\code{\link{get.variety_db}()},
\code{\link{list_db_var}()},
\code{\link{readin_db}()}
}
\concept{access codebook functions}
