% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R,
%   R/factcurator_wrappers.R
\name{list_db_var}
\alias{list_db_var}
\alias{list.db_var}
\title{List the variables (column names) in a given database codebook}
\usage{
list_db_var(db_folder, codebook_name, required_only = FALSE)

list.db_var(db_folder, codebook_name, required_only = FALSE)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{codebook_name}{A string denoting the book name
Use list_db_books() to see options}
}
\description{
List the variables (column names) in a given database codebook

A wrapper for list_db_var to maintain backwards compatibility
}
\seealso{
Other access codebook functions: 
\code{\link{get_variety_db}()},
\code{\link{list_db_books}()},
\code{\link{readin_db}()}
}
\concept{access codebook functions}
