% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_process_fuzzy_module.R
\name{process_fuzzymatch}
\alias{process_fuzzymatch}
\title{Module to process fuzzymatches given an auxiliary file}
\usage{
process_fuzzymatch(output_fuzzymatch_df, aux_fuzzy_status, knitroutputfolder)
}
\arguments{
\item{output_fuzzymatch_df}{A data.frame with the newly generated fuzzymatches}

\item{aux_fuzzy_status}{A data.frame containing the fuzzymatches with is_truematch manually checked}

\item{knitroutputfolder}{A path to send output}
}
\description{
This function writes out:
\itemize{
\item fuzzy_check.csv (A file of fuzzy matches to manually check)
!Note: Rows for which the source column is only "new_fuzzymatch" must be
manually pasted into the aux_fuzzy_status file
Source column containing "csv" denotes that the possible fuzzymatch in
the aux_fuzzy_status file is blank (or not TRUE or FALSE)
}
}
\details{
\itemize{
\item add_fuzzy_to_cv_rename.csv (A file to paste into the main
controlled_vocab/cv_rename.csv)
These are names for which is_truematch = TRUE and these names
should be manually added to controlled_vocab/cv_rename.csv file so they are
caught using do_exactmatch()
\item standardize_new_names.csv (A file of names to standardize for
adding to the database.)
}
}
\seealso{
Other match variety modules: 
\code{\link{do_exactmatch}()},
\code{\link{find_fuzzymatch}()},
\code{\link{process_std_new_names}()}
}
\concept{match variety modules}
