% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R
\name{readin_db}
\alias{readin_db}
\title{Read in all control variable codebooks and create a combined cultivar file}
\usage{
readin_db(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
Read in all control variable codebooks and create a combined cultivar file
}
\seealso{
\code{\link{readin.db}}

Other access codebook functions: 
\code{\link{get.variety_db}()},
\code{\link{list.db_books}()},
\code{\link{list.db_var}()}
}
\concept{access codebook functions}
