% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R
\name{readin_db_init}
\alias{readin_db_init}
\title{Read in database control variables}
\usage{
readin_db_init(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
Note: See read_delim if need to escape backslashes
}
\keyword{internal}
