% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront_data.R
\name{return_validate_message}
\alias{return_validate_message}
\title{Return a message regarding the status of the validation}
\usage{
return_validate_message(confront_summary)
}
\arguments{
\item{output_confront}{An object that is returned by validate::confront()}
}
\description{
Return a message regarding the status of the validation
}
