% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{rm.alias_dupmatch}
\alias{rm.alias_dupmatch}
\title{Remove alias if there is a match with variety
If a variety is listed in the raw data with an alias,
both names may match to names in the database.
This function will remove the alias and return the listed variety name}
\usage{
rm.alias_dupmatch(match_df)
}
\arguments{
\item{match_df}{A data.frame that contains var_id, db_id, and type_db ("alias" or "variety")}
}
\description{
Remove alias if there is a match with variety
If a variety is listed in the raw data with an alias,
both names may match to names in the database.
This function will remove the alias and return the listed variety name
}
