% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_cols_by_cb.R
\name{standardize_cols_by_cb}
\alias{standardize_cols_by_cb}
\title{Standardize the columns in the order specified by the codebook}
\usage{
standardize_cols_by_cb(
  df,
  codebook_name,
  add_missing_cols,
  required_only,
  codebook_cols_only,
  db_folder,
  new_col_fill = ""
)
}
\arguments{
\item{df}{A data.frame with column names to verify against the codebook}

\item{add_missing_cols}{A logical denoting whether to add missing columns}

\item{required_only}{A logical denoting whether to add only required columns.
This parameter is only called if add_missing_cols = TRUE.
Note: This option refers only to missing columns
If optional columns exist in the df, these columns are retained.}

\item{codebook_cols_only}{A logical denoting whether to remove any columns
not specified by the codebook}

\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{new_col_fill}{A value that will be used to fill newly created columns}
}
\description{
Contains options to add missing columns or remove unnecessary columns
The ordering follows the column order specified in the codebook
Any extra columns will come after, in alphabetical order.
}
\concept{standardize to codebook functions}
