% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_fns.R
\name{stringdist.variety}
\alias{stringdist.variety}
\title{Fuzzymatch varieties using stringdist_join}
\usage{
stringdist.variety(
  raw_data,
  database,
  intid_col = "name",
  best_n,
  method_stringdist
)
}
\arguments{
\item{raw_data}{A data.frame}

\item{database}{A data.frame}

\item{intid_col}{A string denoting the name of the internal id column
used to join the raw data and database}

\item{best_n}{A numeric denoting the number of best matches to select}

\item{method_stringdist}{A string corresponding to a string metric defined by stringdist
e.g., "lw", "jw". See \code{\link[stringdist]{stringdist-metrics}} for details.}
}
\description{
Fuzzymatch the varieties from the raw data with the database using an internal id.
This function uses stringdist_join and returns the closest match (smallest distance)
To have enough flexibility without returning too many fuzzy matches,
this function returns the closest match measured in 2 different ways:
\enumerate{
\item select the best_n varieties with the smallest distance, and then select the
best_n varieties that have the smallest scaled distance
\item select the best_n varieties with the smallest scaled distance, irrespective
of the raw distance
See the effect of these different selection methods in test_match_variety_fns.R
}
}
\keyword{internal}
