% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_final_matches_module.R
\name{collect_final_matches}
\alias{collect_final_matches}
\title{Module to collect all the matches from each stage of the variety matching process}
\usage{
collect_final_matches(
  match_list,
  names_raw,
  knitroutputfolder,
  csv_suffix = NULL,
  is_blends = FALSE
)
}
\arguments{
\item{match_list}{A list of the output from each step of the variety matching process,
i.e., exact, rename, and fuzzy. These must be enclosed in a \code{list()} call.}

\item{names_raw}{A data.frame of the names created by \code{create_intid()}}

\item{knitroutputfolder}{A path to the output location}

\item{csv_suffix}{Optional string denoting the suffix for the .csv name.}

\item{is_blends}{A logical that specifies whether the varieties are blends.
Default is FALSE}
}
\description{
\code{collect_final_matches()} combines the match output lists and returns a data frame
containing the raw names, matched clean names, and the stage where they were matched.
An output file, 'final_matches.csv', will be written to the specified output location.
}
\details{
Note: If no match has been found, then the returned match is the original raw name.
}
\seealso{
Other match variety functions: 
\code{\link{create_intid}()},
\code{\link{find_entries_raw_names}()}
}
\concept{match variety functions}
