% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intid.R
\name{create_intid}
\alias{create_intid}
\title{Create internal id for variety names}
\usage{
create_intid(df, variety_col_name, sep_aliases = NULL, ..., alias_col = NULL)
}
\arguments{
\item{df}{A data.frame to with a column containing variety names}

\item{variety_col_name}{A column name denoting the column containing varieties}

\item{sep_aliases}{A regex corresponding to the characters that are used to separate variety aliases}

\item{alias_col}{A column name denoting the column containing aliases}
}
\description{
Note: This function handles aliases passed in as a separate column or aliases
in the same column as the variety name.  Aliases in the same column must be
detectable by a regex provided in sep_aliases
}
\seealso{
Other match variety functions: 
\code{\link{collect_final_matches}()},
\code{\link{find_entries_raw_names}()}
}
\concept{match variety functions}
