% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_exact_module.R
\name{do_exactmatch}
\alias{do_exactmatch}
\title{Do the steps in the exact match}
\usage{
do_exactmatch(
  db_folder,
  data_intid,
  select_before = "2021-01-01",
  select_crops = NULL,
  match_type = "raw",
  rename_df = FALSE
)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{data_intid}{A data.frame created by the function create_intid}

\item{select_before}{A string in the format of Ymd.  The function returns
cultivars that were added to the datebase before this specified date.}

\item{select_crops}{A regular expression of crop types separated by |}

\item{match_type}{A string, either "raw" or "db, " which denotes the type of
matching procedure to use.  If matching raw variety names, use "raw."
If matching database names, use "db."  The difference is that "raw" groups by
the var_id and is only looking for one match per var_id.  The method "db"
groups by the intid and is looking to match every entry provided.}

\item{rename_df}{A logical denoting whether to check the rename (misspelling) file
Otherwise, the matches are pulled from the database list}

\item{data_suffix}{A character suffix to be appended to the columns related to the original data}
}
\description{
Do the steps in the exact match
}
\concept{match variety modules}
