% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_exact_module.R
\name{get_cultivar_rename}
\alias{get_cultivar_rename}
\title{Get cultivar rename list to match.  The cv_rename.csv is located
in the controlled vocabularies.}
\usage{
get_cultivar_rename(rename_df_path)
}
\arguments{
\item{rename_df_path}{The path of the file that contains the variety misspellings}
}
\description{
Get cultivar rename list to match.  The cv_rename.csv is located
in the controlled vocabularies.
}
\keyword{internal}
