% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R,
%   R/factcurator_wrappers.R
\name{list_db_var}
\alias{list_db_var}
\alias{list.db_var}
\title{List the variables (column names) in a given database codebook}
\usage{
list_db_var(db_folder, codebook_name, required_only = FALSE, crop_types)

list.db_var(db_folder, codebook_name, required_only = FALSE)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{codebook_name}{A string denoting the book name.
Use \code{list_db_books()} to see options}

\item{required_only}{A logical, FALSE returns all columns; TRUE returns required columns only}

\item{crop_types}{A vector containing the crop_types.
This is used to select the appropriate traits if codebook_name == "trial_data"}
}
\description{
List the variables (column names) in a given database codebook

A wrapper for list_db_var to maintain backwards compatibility
}
\seealso{
Other access codebook functions: 
\code{\link{get_col_index}()},
\code{\link{get_variety_db}()},
\code{\link{list_db_books}()},
\code{\link{readin_db}()}
}
\concept{access codebook functions}
