% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_final_matches_module.R
\name{match_newdb}
\alias{match_newdb}
\title{Match names with the new db entries
Note: It is important that the "id_newvar" column refers to the same
entries between the output created by create_new_db_entries() and the auxiliary
new cultivar controlled vocabulary names.}
\usage{
match_newdb(auxiliary_files, filename, names_newvarid, ...)
}
\arguments{
\item{filename}{A filename that contains the information
created by create_new_db_entries() and checked and filled in by the collaborators}

\item{names_newvarid}{A data.frame of variety names that do not have matches.  This data.frame
is a list element named 'names_newvarid' generated using create_newdbnames()}

\item{dots}{Bare variables to select from the new cultivar name}
}
\description{
Match names with the new db entries
Note: It is important that the "id_newvar" column refers to the same
entries between the output created by create_new_db_entries() and the auxiliary
new cultivar controlled vocabulary names.
}
