% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_variety_process_fuzzy_module.R
\name{process_fuzzymatch}
\alias{process_fuzzymatch}
\title{Module to process fuzzymatches given an auxiliary file
This function writes out:
\itemize{
\item fuzzy_check.csv (A file of fuzzy matches to manually check)
\item add_fuzzy_to_cv_rename.csv (A file to paste into the main cv_rename.csv)
\item standardize_new_names.csv (A file of names to standardize for
adding to the database.)
}}
\usage{
process_fuzzymatch(output_fuzzymatch_df, aux_fuzzy_status, knitroutputfolder)
}
\arguments{
\item{output_fuzzymatch_df}{A data.frame with the newly generated fuzzymatches}

\item{aux_fuzzy_status}{A data.frame containing the fuzzymatches with is_truematch manually checked}

\item{knitroutputfolder}{A path to send output}
}
\description{
Module to process fuzzymatches given an auxiliary file
This function writes out:
\itemize{
\item fuzzy_check.csv (A file of fuzzy matches to manually check)
\item add_fuzzy_to_cv_rename.csv (A file to paste into the main cv_rename.csv)
\item standardize_new_names.csv (A file of names to standardize for
adding to the database.)
}
}
