% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R
\name{readin.db}
\alias{readin.db}
\title{Read in database control variables}
\usage{
readin.db(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
Note: See read_delim if need to escape backslashes
}
