% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factcurator_wrappers.R
\name{readin.db}
\alias{readin.db}
\title{A wrapper for readin_db to maintain backwards compatibility}
\usage{
readin.db(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
A wrapper for readin_db to maintain backwards compatibility
}
\seealso{
\code{\link{readin_db}}
}
\concept{function rename wrappers}
