% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_controlled_vocab.R,
%   R/factcurator_wrappers.R
\name{readin_db}
\alias{readin_db}
\alias{readin.db}
\title{Read in all control variable codebooks and create a combined cultivar file}
\usage{
readin_db(db_folder)

readin.db(db_folder)
}
\arguments{
\item{db_folder}{A string path to the database controlled vocabulary folder}
}
\description{
Read in all control variable codebooks and create a combined cultivar file

A wrapper for readin_db to maintain backwards compatibility
}
\seealso{
Other access codebook functions: 
\code{\link{get_variety_db}()},
\code{\link{list_db_books}()},
\code{\link{list_db_var}()}
}
\concept{access codebook functions}
