% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_colnames.R
\name{validate_colnames}
\alias{validate_colnames}
\title{Validate column names}
\usage{
validate_colnames(df, codebook_name, db_folder)
}
\arguments{
\item{df}{A data.frame with column names to verify against the codebook}

\item{codebook}{A string denoting the name of the codebook to check the column
names against e.g. "trial_data", "trials_metadata"}
}
\description{
Validate column names
}
