% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_colnames.R
\name{validate_colnames}
\alias{validate_colnames}
\title{Validate column names against codebook}
\usage{
validate_colnames(
  df,
  codebook_name,
  db_folder,
  crop_types = NULL,
  cb_name_remove = NULL
)
}
\arguments{
\item{df}{A data.frame with column names to verify against the codebook
names against e.g. "trial_data", "trials_metadata"}

\item{codebook_name}{A string denoting the book name.
Use \code{list_db_books()} to see options}

\item{db_folder}{A string path to the database controlled vocabulary folder}

\item{crop_types}{A vector containing the crop_types.
This is used to select the appropriate traits if codebook_name == "trial_data"}

\item{cb_name_remove}{A string denoting the name of the codebook to remove from the data (df)
For example, if testing "trial_data", remove the names in df that correspond to "trials_metadata"}
}
\description{
Validate column names against codebook
}
\seealso{
Other validation functions: 
\code{\link{confront_data}()},
\code{\link{create_rules}()},
\code{\link{find_col_info}()}
}
\concept{validation functions}
